refactor helpers methods
