+++ src/reduxTemplates/userform/AddUserForm.js
@@ -15,7 +15,7 @@ import { selectSelectedUser } from "./users.selectors";
 
 import { actions } from "../Users/users.reducer";
 
-import { isEmpty } from "../../helper/isEmpty.js";
+import  isEmpty  from "../../helper";
 
 const { setSelectedUserModal, setSelectedUser } = actions;
 const AddUser = ({ show, handleClose, reset }) => {
